+++ tests/unit/outputs/cli/ci/system/utils/sorting/test_builds.py
@@ -16,46 +16,46 @@
 from unittest import TestCase
 from unittest.mock import Mock
 
-from cibyl.outputs.cli.ci.system.sorting.jobs import SortJobsByName
+from cibyl.outputs.cli.ci.system.utils.sorting.builds import SortBuildsByUUID
 
 
-class TestSortJobsByName(TestCase):
-    """Tests for :class:`SortJobsByName`.
+class TestSortBuildsByUUID(TestCase):
+    """Tests for :class:`SortBuildsByUUID`.
     """
 
-    def test_names_are_equal(self):
-        """Checks that two jobs are equal if they share the same name.
+    def test_uuids_are_equal(self):
+        """Checks that two builds are equal if they share the same uuid.
         """
-        job1 = Mock()
-        job1.name.value = 'job'
+        build1 = Mock()
+        build1.build_id.value = '0'
 
-        job2 = Mock()
-        job2.name.value = 'job'
+        build2 = Mock()
+        build2.build_id.value = '0'
 
-        comparator = SortJobsByName()
+        comparator = SortBuildsByUUID()
 
         self.assertEqual(
             0,
-            comparator.compare(job1, job2)
+            comparator.compare(build1, build2)
         )
 
-    def test_alphabetical_order(self):
-        """Checks that the comparator will sort jobs in alphabetical order.
+    def test_chronological_order(self):
+        """Checks that the comparator will sort builds in chronological order.
         """
-        job1 = Mock()
-        job1.name.value = 'A'
+        build1 = Mock()
+        build1.build_id.value = '9'
 
-        job2 = Mock()
-        job2.name.value = 'B'
+        build2 = Mock()
+        build2.build_id.value = '10'
 
-        comparator = SortJobsByName()
+        comparator = SortBuildsByUUID()
 
         self.assertEqual(
             -1,
-            comparator.compare(job1, job2)
+            comparator.compare(build1, build2)
         )
 
         self.assertEqual(
             1,
-            comparator.compare(job2, job1)
+            comparator.compare(build2, build1)
         )
